requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
openpyxl==3.1.2
msal==1.24.1
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0